+++ package.json
@@ -22,24 +22,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.5",
-    "@angular/common": "^16.2.5",
-    "@angular/compiler": "^16.2.5",
-    "@angular/core": "^16.2.5",
-    "@angular/forms": "^16.2.5",
-    "@angular/platform-browser": "^16.2.5",
-    "@angular/platform-browser-dynamic": "^16.2.5",
-    "@angular/router": "^16.2.5",
+    "@angular/animations": "^17.0.4",
+    "@angular/common": "^17.0.4",
+    "@angular/compiler": "^17.0.4",
+    "@angular/core": "^17.0.4",
+    "@angular/forms": "^17.0.4",
+    "@angular/platform-browser": "^17.0.4",
+    "@angular/platform-browser-dynamic": "^17.0.4",
+    "@angular/router": "^17.0.4",
     "animate.css": "^4.1.1",
     "bootstrap": "^5.2.2",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.2",
-    "@angular/cli": "~16.2.2",
-    "@angular/compiler-cli": "^16.2.5",
+    "@angular-devkit/build-angular": "^17.0.3",
+    "@angular/cli": "~17.0.3",
+    "@angular/compiler-cli": "^17.0.4",
     "@types/jasmine": "~4.3.0",
     "jasmine-core": "~4.5.0",
     "karma": "~6.4.0",
@@ -47,7 +47,8 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "ng-packagr": "^16.2.3",
-    "typescript": "~5.1.6"
+    "ng-packagr": "^17.0.2",
+    "prettier": "^3.1.0",
+    "typescript": "~5.2.2"
   }
 }
