+++ projects/ngx-toast/src/lib/ngx-toast.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { ToastComponent } from './toast/toast.component';
-import { ToastItemComponent } from './toast-item/toast-item.component';
 import { ToastContainerComponent } from './toast-container/toast-container.component';
 import { ToastDirective } from './toast.directive';
 import { CommonModule } from '@angular/common';
@@ -10,22 +9,16 @@ import { ToastIconDirective } from './toast-icon.directive';
 
 
 @NgModule({
-    declarations: [
+    imports: [
+        CommonModule,
         ToastComponent,
-        ToastItemComponent,
         ToastContainerComponent,
         ToastDirective,
         ToastColorDirective,
         ToastIconDirective
     ],
-    imports: [
-        CommonModule
-    ],
     exports: [
-        ToastComponent,
         ToastDirective,
-        ToastColorDirective,
-        ToastIconDirective
     ]
 })
 export class NgxToastModule { }
