--- projects/ngx-toast/src/lib/toast-item/toast-item.component.scss
@@ -1,15 +1,0 @@
-.toast{
-    width: 320px;
-    
-    .toast-icon{
-        // display: block;
-        // // justify-content: center;
-        // // align-items: center;
-        // position: relative;
-        // svg.stack{
-        //     position: absolute;
-        //     top:0;
-        //     left: 0;
-        // }
-    }
-}
