+++ projects/ngx-toast/src/lib/toast.directive.ts
@@ -1,25 +1,14 @@
-import { ComponentFactory, ComponentFactoryResolver, ComponentRef, Directive, OnInit, ViewContainerRef } from '@angular/core';
-import { ToastComponent } from './toast/toast.component';
+import { ComponentRef, Directive, OnInit, ViewContainerRef, inject } from '@angular/core';
+import { ToastContainerComponent } from './toast-container/toast-container.component';
 
 @Directive({
-  selector: '[dagToast]'
+    selector: '[dagToast]',
+    standalone: true
 })
 export class ToastDirective implements OnInit {
-
-    constructor(
-        private vcr: ViewContainerRef,
-        // private cfr: ComponentFactoryResolver
-    ) { }
-
+    private vcr = inject(ViewContainerRef)
 
     ngOnInit() {
-        this.load()
-    }
-
-
-    load() {
-        // let cf: ComponentFactory<ToastComponent> = this.cfr.resolveComponentFactory(ToastComponent)
-        // let cr: ComponentRef<ToastComponent> = this.vcr.createComponent(cf)
-        let cr: ComponentRef<ToastComponent> = this.vcr.createComponent(ToastComponent)
+      const cr: ComponentRef<ToastContainerComponent> = this.vcr.createComponent(ToastContainerComponent)
     }
 }
