+++ projects/ngx-toast/src/lib/toast/toast.component.html
@@ -1,1 +1,109 @@
-<toast-container></toast-container>
+<div
+  class="toast show mb-3 animate__animated"
+  [class.animate__fadeInRight]="animationIn"
+  [class.animate__fadeOutRight]="animationOut"
+  role="alert"
+  aria-live="assertive"
+  aria-atomic="true"
+>
+  <div class="toast-header">
+    <span class="toast-icon me-3">
+      <!-- QUADRATO DI SFONDO COLORATO -->
+      <svg
+        aria-hidden="true"
+        role="img"
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 448 512"
+        width="32px"
+        height="32px"
+        class="stack"
+        [toastColor]="data"
+      >
+        <path
+          fill="currentColor"
+          width="100%"
+          height="100%"
+          d="M400 32H48C21.5 32 0 53.5 0 80v352c0 26.5 21.5 48 48 48h352c26.5 0 48-21.5 48-48V80c0-26.5-21.5-48-48-48z"
+        ></path>
+      </svg>
+
+      <!-- ICONA BIANCA -->
+
+      @switch (data.type) {
+        @case ("error") {
+          <svg
+            width="20px"
+            height="20px"
+            class="stack"
+            role="img"
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 448 512"
+            style="margin-left: -26px"
+          >
+            <path
+              fill="white"
+              d="M439.15 453.06L297.17 384l141.99-69.06c7.9-3.95 11.11-13.56 7.15-21.46L432 264.85c-3.95-7.9-13.56-11.11-21.47-7.16L224 348.41 37.47 257.69c-7.9-3.95-17.51-.75-21.47 7.16L1.69 293.48c-3.95 7.9-.75 17.51 7.15 21.46L150.83 384 8.85 453.06c-7.9 3.95-11.11 13.56-7.15 21.47l14.31 28.63c3.95 7.9 13.56 11.11 21.47 7.15L224 419.59l186.53 90.72c7.9 3.95 17.51.75 21.47-7.15l14.31-28.63c3.95-7.91.74-17.52-7.16-21.47zM150 237.28l-5.48 25.87c-2.67 12.62 5.42 24.85 16.45 24.85h126.08c11.03 0 19.12-12.23 16.45-24.85l-5.5-25.87c41.78-22.41 70-62.75 70-109.28C368 57.31 303.53 0 224 0S80 57.31 80 128c0 46.53 28.22 86.87 70 109.28zM280 112c17.65 0 32 14.35 32 32s-14.35 32-32 32-32-14.35-32-32 14.35-32 32-32zm-112 0c17.65 0 32 14.35 32 32s-14.35 32-32 32-32-14.35-32-32 14.35-32 32-32z"
+            ></path>
+          </svg>
+        }
+        @case ("warning") {
+          <svg
+            class="stack"
+            role="img"
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 576 512"
+            width="20px"
+            height="20px"
+            style="margin-left: -26px"
+          >
+            <path
+              fill="white"
+              d="M569.517 440.013C587.975 472.007 564.806 512 527.94 512H48.054c-36.937 0-59.999-40.055-41.577-71.987L246.423 23.985c18.467-32.009 64.72-31.951 83.154 0l239.94 416.028zM288 354c-25.405 0-46 20.595-46 46s20.595 46 46 46 46-20.595 46-46-20.595-46-46-46zm-43.673-165.346l7.418 136c.347 6.364 5.609 11.346 11.982 11.346h48.546c6.373 0 11.635-4.982 11.982-11.346l7.418-136c.375-6.874-5.098-12.654-11.982-12.654h-63.383c-6.884 0-12.356 5.78-11.981 12.654z"
+            ></path>
+          </svg>
+        }
+        @case ("info") {
+          <svg
+            *ngIf="data.type == 'info'"
+            class="stack"
+            role="img"
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 192 512"
+            width="20px"
+            height="20px"
+            style="margin-left: -26px"
+          >
+            <path
+              fill="white"
+              d="M20 424.229h20V279.771H20c-11.046 0-20-8.954-20-20V212c0-11.046 8.954-20 20-20h112c11.046 0 20 8.954 20 20v212.229h20c11.046 0 20 8.954 20 20V492c0 11.046-8.954 20-20 20H20c-11.046 0-20-8.954-20-20v-47.771c0-11.046 8.954-20 20-20zM96 0C56.235 0 24 32.235 24 72s32.235 72 72 72 72-32.235 72-72S135.764 0 96 0z"
+            ></path>
+          </svg>
+        }
+      }
+    </span>
+
+
+    <strong class="me-auto">{{ data.title }}</strong>
+    <small class="text-muted">{{ data.time | date }}</small>
+    <button
+      type="button"
+      class="ms-2 mb-1 btn-close"
+      (click)="onClose()"
+      aria-label="Close"
+    >
+      <span aria-hidden="true"></span>
+    </button>
+  </div>
+  <div class="toast-body position-relative">
+    <p>
+      {{ data.message }}
+      <small
+        class="position-absolute mb-2 me-3"
+        style="bottom: 0; right: 0"
+        [toastColor]="data"
+      >
+        {{ data.type }}
+      </small>
+    </p>
+  </div>
+</div>
