+++ projects/ngx-toast/src/lib/toast/toast.component.scss
@@ -1,0 +1,3 @@
+.toast {
+  width: 320px;
+}
