+++ projects/ngx-toast/src/public-api.ts
@@ -4,8 +4,8 @@
 
 export * from './lib/ngx-toast.service';
 export * from './lib/ngx-toast.module';
-export * from './lib/toast.directive';
-export * from './lib/toast-icon.directive';
-export * from './lib/toast-color.directive';
-export * from './lib/toast/toast.component';
+// export * from './lib/toast.directive';
+// export * from './lib/toast-icon.directive';
+// export * from './lib/toast-color.directive';
+// export * from './lib/toast/toast.component';
